Move NuGetSourcesTests
